+++ src/api/projects/merge_requests/awards.rs
@@ -8,8 +8,23 @@
 //!
 //! These endpoints are used for querying project merge request awards.
 
+mod award;
 mod awards;
+mod create;
+mod delete;
+
+pub use self::award::MergeRequestAward;
+pub use self::award::MergeRequestAwardBuilder;
+pub use self::award::MergeRequestAwardBuilderError;
 
 pub use self::awards::MergeRequestAwards;
 pub use self::awards::MergeRequestAwardsBuilder;
 pub use self::awards::MergeRequestAwardsBuilderError;
+
+pub use self::create::CreateMergeRequestAward;
+pub use self::create::CreateMergeRequestAwardBuilder;
+pub use self::create::CreateMergeRequestAwardBuilderError;
+
+pub use self::delete::DeleteMergeRequestAward;
+pub use self::delete::DeleteMergeRequestAwardBuilder;
+pub use self::delete::DeleteMergeRequestAwardBuilderError;