+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Additions
 
+  * Add missing endpoints to `api::projects::merge_requests::awards` and `api::projects::merge_requests::notes::awards`.
   * Add `api::projects::issues::awards` and `api::projects::issues::notes::awards` endpoints.
 
 # v0.1606.0